feat: bundle mode integration + relevant cleanup
