+++ example/babel.config.js
@@ -1,6 +1,16 @@
 const path = require('path');
 const pak = require('../package.json');
 
+/** @type {import('react-native-worklets/plugin').PluginOptions} */
+const workletsPluginOptions = {
+  bundleMode: true,
+  strictGlobal: true,
+  importForwarding: {
+    moduleNames: ['expensify-common', '@expensify/react-native-live-markdown'],
+    relativePaths: ['react-native-live-markdown/src/'],
+  },
+};
+
 module.exports = {
   presets: ['module:@react-native/babel-preset'],
   plugins: [
@@ -13,6 +23,6 @@ module.exports = {
         },
       },
     ],
-    'react-native-worklets/plugin',
+    ['react-native-worklets/plugin', workletsPluginOptions],
   ],
 };
