+++ example/metro.config.js
@@ -1,4 +1,5 @@
 const {getDefaultConfig, mergeConfig} = require('@react-native/metro-config');
+const {bundleModeMetroConfig} = require('react-native-worklets/bundleMode');
 const path = require('path');
 
 const root = path.resolve(__dirname, '..');
@@ -22,4 +23,8 @@ const config = {
   },
 };
 
-module.exports = mergeConfig(getDefaultConfig(__dirname), config);
+module.exports = mergeConfig(
+  getDefaultConfig(__dirname),
+  bundleModeMetroConfig,
+  config,
+);
