+++ patches/react-native++metro-runtime+0.84.4.patch
@@ -1,0 +1,36 @@
+diff --git a/node_modules/react-native/node_modules/metro-runtime/src/modules/HMRClient.js b/node_modules/react-native/node_modules/metro-runtime/src/modules/HMRClient.js
+index 534ac87..5773985 100644
+--- a/node_modules/react-native/node_modules/metro-runtime/src/modules/HMRClient.js
++++ b/node_modules/react-native/node_modules/metro-runtime/src/modules/HMRClient.js
+@@ -3,6 +3,9 @@
+ const EventEmitter = require("./vendor/eventemitter3");
+ const HEARTBEAT_INTERVAL_MS = 20_000;
+ const inject = ({ module: [id, code], sourceURL }) => {
++  if (global.__workletsModuleProxy?.propagateModuleUpdate) {
++    global.__workletsModuleProxy.propagateModuleUpdate(code, sourceURL);
++  }
+   if (global.globalEvalWithSourceUrl) {
+     global.globalEvalWithSourceUrl(code, sourceURL);
+   } else {
+diff --git a/node_modules/react-native/node_modules/metro-runtime/src/polyfills/require.js b/node_modules/react-native/node_modules/metro-runtime/src/polyfills/require.js
+index c46fb7c..2ba29ab 100644
+--- a/node_modules/react-native/node_modules/metro-runtime/src/polyfills/require.js
++++ b/node_modules/react-native/node_modules/metro-runtime/src/polyfills/require.js
+@@ -334,14 +334,14 @@ function unknownModuleError(id) {
+   }
+   return Error(message);
+ }
++metroRequire.getModules = () => {
++  return modules;
++};
+ if (__DEV__) {
+   metroRequire.Systrace = {
+     beginEvent: () => {},
+     endEvent: () => {},
+   };
+-  metroRequire.getModules = () => {
+-    return modules;
+-  };
+   var createHotReloadingObject = function () {
+     const hot = {
+       _acceptCallback: null,
