+++ src/__tests__/webParser.test.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import {expect} from '@jest/globals';
 import {parseRangesToHTMLNodes} from '../web/utils/parserUtils';
-import parseExpensiMark from '../parseExpensiMark';
+import parseExpensiMark from '../parse';
 
 declare module 'expect' {
   interface Matchers<R> {
