+++ src/index.tsx
@@ -1,4 +1,4 @@
 export {default as MarkdownTextInput, getWorkletRuntime} from './MarkdownTextInput';
 export type {MarkdownTextInputProps, MarkdownStyle} from './MarkdownTextInput';
 export type {MarkdownType, MarkdownRange} from './commonTypes';
-export {default as parseExpensiMark} from './parseExpensiMark';
+export {default as parseExpensiMark} from './parse';
